*Move vector types into own header -Remove common header
